Refactor files tests and core tests
