+++ tests/unit/core/test_profile.py
@@ -1,0 +1,20 @@
+"""Unit tests for the Zowe Python SDK Core package."""
+
+import os
+import unittest
+
+from zowe.core_for_zowe_sdk import (
+    ZosmfProfile
+)
+
+class TestZosmfProfileClass(unittest.TestCase):
+    """ZosmfProfile class unit tests."""
+
+    def setUp(self):
+        """Setup fixtures for ZosmfProfile class."""
+        self.profile_name = "MOCK"
+
+    def test_object_should_be_instance_of_class(self):
+        """Created object should be instance of ZosmfProfile class."""
+        zosmf_profile = ZosmfProfile(self.profile_name)
+        self.assertIsInstance(zosmf_profile, ZosmfProfile)
